2007-07-27  Jorn Baayen  <jorn@openedhand.com>
	* libgssdp/gssdp-client.c: (socket_source_cb):

	Fix compiler warning. Fixes #421.

	Patch by Priit Laes.



git-svn-id: https://svn.o-hand.com/repos/gupnp/trunk/gssdp@343 d8cb91d7-bff9-0310-92b9-80b65e4482b2

+++ libgssdp/gssdp-client.c
@@ -513,9 +513,9 @@ socket_source_cb (gpointer user_data)
                           &addr_size);
 
         if (bytes >= BUF_SIZE) {
-                g_warning ("Received packet of %d bytes, but the maximum "
+                g_warning ("Received packet of %u bytes, but the maximum "
                            "buffer size is %d. Packed dropped.",
-                           bytes, BUF_SIZE);
+                           (unsigned int) bytes, BUF_SIZE);
 
                 return TRUE;
         }
